# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trivelo Hotel Booking API",
        "description": "Comprehensive hotel booking system API with multi-role authentication",
        "contact": {
            "email": "support@trivelo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/hotels/{id}": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get hotel details",
                "description": "Get detailed information about a specific hotel",
                "operationId": "42ab12b8021f7fe329cd004cb4b277a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hotel ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "hotel": {
                                                    "$ref": "#/components/schemas/Hotel"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hotel not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Hotel": {
                "title": "Hotel",
                "description": "Hotel model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "address",
                    "city",
                    "country",
                    "star_rating",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Hotel ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Hotel owner/manager user ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 2
                    },
                    "name": {
                        "description": "Hotel name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Grand Plaza Hotel"
                    },
                    "description": {
                        "description": "Hotel description",
                        "type": "string",
                        "example": "Luxury hotel in city center",
                        "nullable": true
                    },
                    "email": {
                        "description": "Hotel contact email",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "info@grandplaza.com"
                    },
                    "phone": {
                        "description": "Hotel contact phone",
                        "type": "string",
                        "maxLength": 20,
                        "example": "+1234567890"
                    },
                    "website": {
                        "description": "Hotel website URL",
                        "type": "string",
                        "example": "https://grandplaza.com",
                        "nullable": true
                    },
                    "address": {
                        "description": "Hotel address",
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "city": {
                        "description": "Hotel city",
                        "type": "string",
                        "maxLength": 100,
                        "example": "New York"
                    },
                    "state": {
                        "description": "Hotel state",
                        "type": "string",
                        "maxLength": 100,
                        "example": "NY",
                        "nullable": true
                    },
                    "country": {
                        "description": "Hotel country",
                        "type": "string",
                        "maxLength": 100,
                        "example": "USA"
                    },
                    "postal_code": {
                        "description": "Hotel postal code",
                        "type": "string",
                        "maxLength": 20,
                        "example": "10001",
                        "nullable": true
                    },
                    "latitude": {
                        "description": "Hotel latitude",
                        "type": "number",
                        "format": "float",
                        "example": 40.7128,
                        "nullable": true
                    },
                    "longitude": {
                        "description": "Hotel longitude",
                        "type": "number",
                        "format": "float",
                        "example": -74.006,
                        "nullable": true
                    },
                    "star_rating": {
                        "description": "Hotel star rating",
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 4
                    },
                    "check_in_time": {
                        "description": "Check-in time",
                        "type": "string",
                        "format": "time",
                        "example": "15:00:00"
                    },
                    "check_out_time": {
                        "description": "Check-out time",
                        "type": "string",
                        "format": "time",
                        "example": "11:00:00"
                    },
                    "cancellation_policy": {
                        "description": "Cancellation policy",
                        "type": "string",
                        "example": "Free cancellation up to 24 hours",
                        "nullable": true
                    },
                    "status": {
                        "description": "Hotel status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected",
                            "suspended"
                        ],
                        "example": "approved"
                    },
                    "is_active": {
                        "description": "Hotel active status",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "rooms": {
                        "description": "Hotel rooms",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Room"
                        }
                    },
                    "amenities": {
                        "description": "Hotel amenities",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Amenity"
                        }
                    },
                    "rooms_count": {
                        "description": "Total number of rooms",
                        "type": "integer",
                        "example": 50
                    },
                    "available_rooms_count": {
                        "description": "Number of available rooms",
                        "type": "integer",
                        "example": 15
                    },
                    "average_rating": {
                        "description": "Average rating",
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "reviews_count": {
                        "description": "Total number of reviews",
                        "type": "integer",
                        "example": 125
                    }
                },
                "type": "object"
            },
...
            "Amenity": {
                "title": "Amenity",
                "description": "Hotel/Room amenity model",
                "required": [
                    "id",
                    "name",
                    "type",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Amenity ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Amenity name",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Free WiFi"
                    },
                    "description": {
                        "description": "Amenity description",
                        "type": "string",
                        "example": "High-speed internet access",
                        "nullable": true
                    },
                    "type": {
                        "description": "Amenity type",
                        "type": "string",
                        "enum": [
                            "hotel",
                            "room",
                            "both"
                        ],
                        "example": "both"
                    },
                    "icon": {
                        "description": "Amenity icon",
                        "type": "string",
                        "maxLength": 50,
                        "example": "wifi",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Amenity active status",
                        "type": "boolean",
                        "example": true
                    },
                    "sort_order": {
                        "description": "Sort order",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}